TestUtil: Annotate as @Ignore and remove fake test method
The TestUtil class has a dummy test method, because bazel test
will fail with "No runnable methods" on any test source file
that has no @Test annotated methods.

A better way to avoid this error is to annotate the class as
@Ignore, so that bazel does not even attempt to run tests.

Actually, an even better way would be to split TestUtil out into
a java_libary so it's not included in test sources in the first
place, but that's more work.

Change-Id: I25e99d3aa55e317df90a0af8482e7352db0ab444