Fix the version of postgresql used in the test container
PostgreSQLContainer defines the version of postgresql that will be
used in the container. The version can be updated between versions of
the PostgreSQLContainer, which results in a new image being downloaded
when the tests are run. Over time this can result in multiple unused
images being left on the test system.

Also, PostgreSQLContainer by default uses the regular image which is
rather large. For example the "9.6.12" image is 230MB, compared to the
"9.6.12-alpine" image which is only 37.7MB.

Fix the version to "9.6.12-alpine" to make use of the smaller image,
and prevent futher images from unnecessarily being downloaded if (or
when) PostgreSQLContainer's default image is updated.

Change-Id: I6b804b1a0049ae2a71ea89b4253164fdc1612e0c