Upgrade testcontainers to 1.15.0-rc2
As per [1], "this release includes a fix for a breaking change that
appeared in Docker for Mac v2.4.0.0 (#3159). We recommend upgrading to
this version of Testcontainers ASAP".

This issue was reproduced locally with the currently latest Docker
Desktop version 2.4.0.0 for macOS. This upgrade allows plugin tests to
pass again. An upgrade to the next (non-rc) stable version of
testcontainers will be done once available.

Upgrade slf4j (based on [2]) and postgresql ([3]) versions accordingly.

Add the docker-java dependencies based on [2] as well.

[1] https://github.com/testcontainers/testcontainers-java/releases/tag/1.15.0-rc2
[2] https://search.maven.org/artifact/org.testcontainers/testcontainers/1.15.0-rc2/jar
[3] https://github.com/testcontainers/testcontainers-java/compare/1.14.3...1.15.0-rc2

Change-Id: I33180bc0dbf8f2ccf8850b4af2de6bd85038e59c