Add an executable README that sets gc-executor up
Add an executable README to serve as an example for building, locally
installing and starting the gc-executor. Includes generating gc.config.

The script also copies the available postgresql jar to test site's /lib.

The gc-conductor plugin still has to be installed within that site,
along with a running and provisioned postgresql database. The sibling
documentation should readily explain how to do that.

Change-Id: I210dd1c60f2373aa68ce69e15831f6f953289a4e