Customize Gatling test values to execute garbage collection
Before this change, the default values of the plugin's configuration
were taken as the values used during the test. This isn't the case as
every environment can have its own configuration leading to a failure in
the gatling tests.

This change introduces two optional variables that can be configured in the
gatling test to configure the test according to the system.

The first variable is timemultiplier, this variable multiplies the
default number of seconds which is 60 seconds. If not configured, this
value is set to 1.

The second variable is looseobjects, this variable is used to control
the number of changes used during the test. If not configured, this
variable is set to a default 400.

Change-Id: I81eb81542b4218e71d4d3a42863c9109a3af4af9