CreateChangesTriggeringGc: Customize data to execute garbage collection
Before this change, the default values of the plugin's configuration
were taken as the values used during the test. This isn't the case
anymore, as every environment can have its own configuration by
specifying gc-conductor plugin properties in the gerrit.config file,
leading to a failure in the gatling tests.

This change introduces two optional environment properties to configure
the gatling test according to the system.

The first property is minute_multiplier, which  multiplies the default
number of seconds which is 60 seconds. If not configured, this value is
set to 1.

The second property is loose_objects, used to control the number of
changes used during the test. If not configured, this variable is set to
the plugin's default value, 400.

Change-Id: I81eb81542b4218e71d4d3a42863c9109a3af4af9