Trust user in Gerrit docker setup
Add --dev parameter into Gerrit site initialization command to enable
login with initialized user. Enable trustContainerAuth setting in
Gerrit config to authenticate user over http traffic for git. In this
case, Gerrit trusts any registered user. Initially Gerrit trusts admin
user because it is created by default. It is not requried to generate
http password for the user.
This is useful for automation, when using the docker solution in
gc-conductor plugin, introduced in [1].

[1] https://gerrit-review.googlesource.com/c/plugins/gc-conductor/+/338374

Feature: Issue 16096
Change-Id: Ia8bb7a03661dcab035f4f95d33bb7c8222653552