Remove the local port binding from gerrit docker container
This change removes the port binding from docker containers to local
ports while exposing the docker container port. This removal is because
the port chosen to be bind to can be in use by the local machine.
Therefore, docker will refuse to create the container.

As a result of this change, gerrit docker container can be reffered by
the ip address assign to it instead of localhost acting as a reverse
proxy.

Change-Id: I370241ee48153712243c1c1dbce6b20e77eed815